+++ src/stemmer.rs
@@ -16,7 +16,7 @@ type SbSymbol = u8;
 ///
 /// ```
 /// use stemmer::Stemmer;
-/// let stemmer = Stemmer::new("french").unwrap();
+/// let mut stemmer = Stemmer::new("french").unwrap();
 /// assert_eq!("éternu", &stemmer.stem("éternuerai"));
 /// ```
 pub struct Stemmer {
@@ -108,30 +108,24 @@ impl Stemmer {
     /// Stems the `word`.
     ///
     /// Returns an owned string.
-    pub fn stem(&self, word: &str) -> String {
-        unsafe {
-            (&*self.stem_unsafe(word)).to_string()
-        }
+    pub fn stem(&mut self, word: &str) -> String {
+        String::from(self.stem_str(word))
     }
 
-    /// Stems the word, unsafely.
+    /// Stems the word and returns a str
     ///
-    /// The `str` pointer it returns is only valid as long as you don't
+    /// The `str` reference it returns is only valid as long as you don't
     /// call `stem` or stem_unsafe` again.
     ///
-    /// Unless you know what you are doing, you should use the safe `stem` method.
-    ///
-    /// (Note that this method is actually safe to call, but derefencing the pointer it
-    /// returns is unsafe)
-    pub fn stem_unsafe (&self, word: &str) -> *const str {
+    /// You should use the `stem` method instead.
+    pub fn stem_str (&mut self, word: &str) -> &str {
         unsafe {
             let word = CString::new(word).unwrap();
             let res = sb_stemmer_stem(self.stemmer,
                                       word.as_ptr(),
                                       word.to_bytes().len() as i32);
             let bytes:&[u8] = CStr::from_ptr(res).to_bytes();
-            let s:&str = str::from_utf8_unchecked(bytes);
-            s
+            str::from_utf8_unchecked(bytes)
         }
     }
 }